Support NSUserDefaults with suiteName
